feat(getParamNames): support async with single param
Bug was found on async function with single parameter and no parenthesis.
Although it could be fixed by enriching the regular expression:

if (/^(?:async)?\s*[^(]+?\s*=>/.test(declaration)) {
    // covers the following cases: "name =>", "async a =>"
    return [declaration.match(/^(?:async)?\s*(\S+)\s*=>/)[1]]
  }

a better strategy is to use a JavaScript parser.
Acorn is the smallest (and most maintained) one.

Although use standardjs eslint configuration.